dynamically choose lemma char copy option
Use the lemma char copy option that results in the shortest lemma script
list.

This is broken for annotation schema merging, meaning that fine-tuning
will not work right. This doesn't seem to make any difference for
English, so I'm abandoning it. Should probably test it elsewhere as
well, but oh well. If we notice horrible lemma performance in the
future, we can revisit.