Fix intermittent test failures for multi-ling pred
This took a TON of time to track down. Here's what happened:
- training is true by default when instantiating a torch module
- the tagger layer contains DropOut, which is non-deterministic and also
differs at train and test time
- when we re-instantiate the tagger, it's set to training by default.
- the training param is not part of the state_dict I guess? Python can
tell it's different, but DiffLib died trying to compare (iteration over
0d Tensor) or else I guess it probably would have found the issue.

Call `train()` on the tagger layer after instantiating it so that its
`training` value matches that of the parent module. Also add an `eval`
call in the predictor for good measure.